# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a fundoo notes application.You can create notes here",
        "version": "1.0.0",
        "title": "Fundoo notes app",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
...
    "paths" : {
...
    "/users/login": {
        "post": {
          "tags": [
            "user"
          ],
          "summary": "Logs user into the system",
          "description": "",
          "operationId": "loginUser",
          "produces": [
            "application/xml",
            "application/json"
          ],
          "parameters": [
            {
                "in": "body",
                "name": "body",
                "description": "login credentials",
                "required": true,
                "schema": {
                  "$ref": "#/loginUser"
                }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "string"
              }
            },
            "400": {
              "description": "Invalid email/password supplied"
            }
          }
        }
    }
  }
}